Update AgeRestriction.java
Marian’s library is facing a major epidemic. Some of the local schoolchildren have been checking
out adult books that are entirely inappropriate for their age. These books contain topics and
illustrations that might scare children. In light of this, Marian needs you to write a program to help figure out which books a given person is allowed to borrow.